Fix crash in Apportable because isKindOfClass:[NSMutableDictionary class] isn't a good check for mutability
The reason this is necessary is because the implementation of __NSCFDictionary in Apportable (the class dictionary literals end up being) is a subclass
of NSMutableDictionary. This is the same as in iOS, but the iOS runtime doesn't use __NSCFDictionary almost ever, so PubNub gets away with this. Checking
for isKindOfClass:[NSMutableDictionary class] is just a bad idea, so I changed it in PubNub.